+++ plugins/source/azuredevops/resources/plugin/tables.go
@@ -1,0 +1,10 @@
+package plugin
+
+import (
+	"github.com/cloudquery/cloudquery/plugins/source/azuredevops/resources/services/core"
+	"github.com/cloudquery/plugin-sdk/schema"
+)
+
+func tables() []*schema.Table {
+	return []*schema.Table{core.Projects()}
+}