+++ plugins/source/gcp/codegen/recipes/serviceusage.go
@@ -7,8 +7,6 @@ import (
 	pb "google.golang.org/genproto/googleapis/api/serviceusage/v1"
 )
 
-
-
 func init() {
 	resources := []*Resource{
 		{
@@ -37,7 +35,8 @@ func init() {
 		resource.Template = "newapi_list"
 		resource.MockTemplate = "newapi_list_grpc_mock"
 		resource.RequestStructFields = `Parent: "projects/" + c.ProjectId,
-		PageSize: 200,`
+		PageSize: 200,
+		Filter: "state:ENABLED",`
 	}
 
 	Resources = append(Resources, resources...)