+++ plugins/source/gcp/resources/services/serviceusage/services.go
@@ -56,12 +56,13 @@ func fetchServices(ctx context.Context, meta schema.ClientMeta, parent *schema.R
 	req := &pb.ListServicesRequest{
 		Parent:   "projects/" + c.ProjectId,
 		PageSize: 200,
+		Filter:   "state:ENABLED",
 	}
 	gcpClient, err := serviceusage.NewClient(ctx, c.ClientOptions...)
 	if err != nil {
 		return err
 	}
-	it := gcpClient.ListServices(ctx, req)
+	it := gcpClient.ListServices(ctx, req, c.CallOptions...)
 	for {
 		resp, err := it.Next()
 		if err == iterator.Done {